+++ CHANGELOG.md
@@ -9,4 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.0.1] 2018-03-13
 
-- initial release with scaffold
+- Initial release with scaffold
+- Inclued some features:
+  - Setup VI Server with Credential
+  - Pass cmdlet directly
+  - Snapshot commands